Resolve conflicts in Authorizer (keep PAIMON_VIEW on the VIEW
privilege path, take FLUSS from main) and RangerInterfaceTest imports.

+++ fe/fe-core/src/main/java/com/starrocks/catalog/PaimonTable.java
@@ -51,7 +51,7 @@ public class PaimonTable extends Table {
 
     public PaimonTable(String catalogName, String dbName, String tblName, List<Column> schema,
                        org.apache.paimon.table.Table paimonNativeTable) {
-        super(CONNECTOR_ID_GENERATOR.getNextId().asInt(), tblName, TableType.PAIMON, schema);
+        super(CONNECTOR_ID_GENERATOR.getNextId().asLong(), tblName, TableType.PAIMON, schema);
         this.catalogName = catalogName;
         this.databaseName = dbName;
         this.tableName = tblName;
